Using animated Sprites in Vbox
This is done very poorly but just a proof of concept that you can have the animated sprite in the vbox container. Basically I made an Node2D which has a script which extends base enemy class, the enemy stats resource is an export variable for this class (for actual game will probs need to use a setter function). This enemy is then a child of a control node (only control nodes get listed properly in box containers)